requirements.txt
# Python dependencies for Confluence Translations
# Python 3.8+ required

# Translation APIs - Choose one or more
deepl>=1.15.0
google-cloud-translate>=3.15.0

# Environment variable management
python-dotenv>=1.0.0

# YAML support for config files
pyyaml>=6.0

# REST API server (optional, for remote database access)
# Uncomment if you need remote database access via API:
# flask>=2.0.0
# flask-cors>=3.0.0

